Changed to load all then process
