+++ sdk/appconfiguration/azure-appconfiguration-provider/azure/appconfiguration/provider/_azureappconfigurationproviderbase.py
@@ -3,6 +3,9 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # -------------------------------------------------------------------------
+import base64
+import hashlib
+import json
 import os
 import random
 import time
@@ -25,6 +28,9 @@ from typing import (
     ValuesView,
     TypeVar,
 )
+from azure.appconfiguration import (  # type:ignore # pylint:disable=no-name-in-module
+    FeatureFlagConfigurationSetting,
+)
 from ._models import SettingSelector
 from ._constants import (
     REQUEST_TRACING_DISABLED_ENVIRONMENT_VARIABLE,
@@ -38,8 +44,15 @@ from ._constants import (
     PERCENTAGE_FILTER_KEY,
     TIME_WINDOW_FILTER_KEY,
     TARGETING_FILTER_KEY,
+    PERCENTAGE_FILTER_NAMES,
+    TIME_WINDOW_FILTER_NAMES,
+    TARGETING_FILTER_NAMES,
+    TELEMETRY_KEY,
+    METADATA_KEY,
+    ETAG_KEY,
+    FEATURE_FLAG_REFERENCE_KEY,
+    ALLOCATION_ID_KEY,
 )
-
 JSON = Mapping[str, Any]
 _T = TypeVar("_T")
 logger = logging.getLogger(__name__)
@@ -260,7 +273,7 @@ class AzureAppConfigurationProviderBase(Mapping[str, Union[str, JSON]]):  # pyli
     """
 
     def __init__(self, **kwargs: Any) -> None:
-        self._origin_endpoint = kwargs.get("endpoint", None)
+        self._origin_endpoint: str = kwargs.get("endpoint", "")
         self._dict: Dict[str, Any] = {}
         self._selects: List[SettingSelector] = kwargs.pop(
             "selects", [SettingSelector(key_filter="*", label_filter=NULL_CHAR)]
@@ -282,10 +295,10 @@ class AzureAppConfigurationProviderBase(Mapping[str, Union[str, JSON]]):  # pyli
         )
         self._feature_flag_enabled = kwargs.pop("feature_flag_enabled", False)
         self._feature_flag_selectors = kwargs.pop("feature_flag_selectors", [SettingSelector(key_filter="*")])
-        self._refresh_on_feature_flags: Mapping[Tuple[str, str], Optional[str]] = {}
+        self._refresh_on_feature_flags: Dict[Tuple[str, str], Optional[str]] = {}
         self._feature_flag_refresh_timer: _RefreshTimer = _RefreshTimer(**kwargs)
         self._feature_flag_refresh_enabled = kwargs.pop("feature_flag_refresh_enabled", False)
-        self._feature_filter_usage: Mapping[str, bool] = {}
+        self._feature_filter_usage: Dict[str, bool] = {}
         self._uses_load_balancing = kwargs.pop("load_balancing_enabled", False)
         self._uses_ai_configuration = False
         self._uses_aicc_configuration = False  # AI Chat Completion
@@ -300,6 +313,120 @@ class AzureAppConfigurationProviderBase(Mapping[str, Union[str, JSON]]):  # pyli
                 trimmed_key = config.key[len(trim) :]
                 break
         return trimmed_key
+
+    def _feature_flag_telemetry(
+        self, endpoint: str, feature_flag: FeatureFlagConfigurationSetting, feature_flag_value: Dict
+    ):
+        if TELEMETRY_KEY in feature_flag_value:
+            if METADATA_KEY not in feature_flag_value[TELEMETRY_KEY]:
+                feature_flag_value[TELEMETRY_KEY][METADATA_KEY] = {}
+            feature_flag_value[TELEMETRY_KEY][METADATA_KEY][ETAG_KEY] = feature_flag.etag
+
+            if not endpoint.endswith("/"):
+                endpoint += "/"
+            feature_flag_reference = f"{endpoint}kv/{feature_flag.key}"
+            if feature_flag.label and not feature_flag.label.isspace():
+                feature_flag_reference += f"?label={feature_flag.label}"
+            if feature_flag_value[TELEMETRY_KEY].get("enabled"):
+                feature_flag_value[TELEMETRY_KEY][METADATA_KEY][FEATURE_FLAG_REFERENCE_KEY] = feature_flag_reference
+                allocation_id = self._generate_allocation_id(feature_flag_value)
+                if allocation_id:
+                    feature_flag_value[TELEMETRY_KEY][METADATA_KEY][ALLOCATION_ID_KEY] = allocation_id
+
+    def _feature_flag_appconfig_telemetry(self, feature_flag: FeatureFlagConfigurationSetting):
+        if feature_flag.filters:
+            for filter in feature_flag.filters:
+                if filter.get("name") in PERCENTAGE_FILTER_NAMES:
+                    self._feature_filter_usage[PERCENTAGE_FILTER_KEY] = True
+                elif filter.get("name") in TIME_WINDOW_FILTER_NAMES:
+                    self._feature_filter_usage[TIME_WINDOW_FILTER_KEY] = True
+                elif filter.get("name") in TARGETING_FILTER_NAMES:
+                    self._feature_filter_usage[TARGETING_FILTER_KEY] = True
+                else:
+                    self._feature_filter_usage[CUSTOM_FILTER_KEY] = True
+
+    @staticmethod
+    def _generate_allocation_id(feature_flag_value: Dict[str, JSON]) -> Optional[str]:
+        """
+        Generates an allocation ID for the specified feature.
+        seed=123abc\ndefault_when_enabled=Control\npercentiles=0,Control,20;20,Test,100\nvariants=Control,standard;Test,special # pylint:disable=line-too-long
+
+        :param  Dict[str, JSON] feature_flag_value: The feature to generate an allocation ID for.
+        :rtype: str
+        :return: The allocation ID.
+        """
+
+        allocation_id = ""
+        allocated_variants = []
+
+        allocation: Optional[JSON] = feature_flag_value.get("allocation")
+
+        if not allocation:
+            return None
+
+        # Seed
+        allocation_id = f"seed={allocation.get('seed', '')}"
+
+        # DefaultWhenEnabled
+        if "default_when_enabled" in allocation:
+            allocated_variants.append(allocation.get("default_when_enabled"))
+
+        allocation_id += f"\ndefault_when_enabled={allocation.get('default_when_enabled', '')}"
+
+        # Percentile
+        allocation_id += "\npercentiles="
+
+        percentile = allocation.get("percentile")
+
+        if percentile:
+            percentile_allocations = sorted(
+                (x for x in percentile if x.get("from") != x.get("to")),
+                key=lambda x: x.get("from"),
+            )
+
+            for percentile_allocation in percentile_allocations:
+                if "variant" in percentile_allocation:
+                    allocated_variants.append(percentile_allocation.get("variant"))
+
+            allocation_id += ";".join(
+                f"{pa.get('from')}," f"{base64.b64encode(pa.get('variant').encode()).decode()}," f"{pa.get('to')}"
+                for pa in percentile_allocations
+            )
+
+        if not allocated_variants and not allocation.get("seed"):
+            return None
+
+        # Variants
+        allocation_id += "\nvariants="
+
+        variants_value = feature_flag_value.get("variants")
+        if variants_value and (isinstance(variants_value, list) or all(isinstance(v, dict) for v in variants_value)):
+            if (
+                allocated_variants
+                and isinstance(variants_value, list)
+                and all(isinstance(v, dict) for v in variants_value)
+            ):
+                sorted_variants: List[Dict[str, Any]] = sorted(
+                    (v for v in variants_value if v.get("name") in allocated_variants),
+                    key=lambda v: v.get("name"),
+                )
+
+                for v in sorted_variants:
+                    allocation_id += f"{base64.b64encode(v.get('name', '').encode()).decode()},"
+                    if "configuration_value" in v:
+                        allocation_id += (
+                            f"{json.dumps(v.get('configuration_value', ''), separators=(',', ':'), sort_keys=True)}"
+                        )
+                    allocation_id += ";"
+                if sorted_variants:
+                    allocation_id = allocation_id[:-1]
+
+        # Create a sha256 hash of the allocation_id
+        hash_object = hashlib.sha256(allocation_id.encode())
+        hash_digest = hash_object.digest()
+
+        # Encode the first 15 bytes in base64 url
+        return base64.urlsafe_b64encode(hash_digest[:15]).decode()
 
     def __getitem__(self, key: str) -> Any:
         # pylint:disable=docstring-missing-param,docstring-missing-return,docstring-missing-rtype
